Moved all graphics component into seperate window class. Full code was just ported over and I am shocked that this worked
