+++ SolSim/SolSim.vcxproj
@@ -142,9 +142,11 @@
     <ClCompile Include="glad.c" />
     <ClCompile Include="SolSim.cpp" />
     <ClCompile Include="Logger.cpp" />
+    <ClCompile Include="window.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Logger.h" />
+    <ClInclude Include="window.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
